Fix -Wformat-security warning for variadic templates (#1260)
Summary:
If -Wformat-security is enabled,
rdb_fatal_error(... "not a printf format string")
gives
rdb_utils.h:254:27: error: format string is not a string literal (potentially insecure) [-Werror,-Wformat-security]
  myrocks_sql_print_error(fmt, std::forward<Params>(params)...);

Fix by introducing an rdb_fatal_error overload for messages without format strings, i.e. the ones which could serve as a parameter for a "%s" format string.

At the same time at a format function attribute for rdb_fatal_error for supported compilers, which at the moment are clang >= 15.

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1260

Reviewed By: hermanlee

Differential Revision: D42527016

Pulled By: yoshinorim

fbshipit-source-id: 094f7884a043d26dae7737b285b505e87519875b

+++ storage/rocksdb/rdb_datadic.cc
@@ -5489,11 +5489,13 @@ bool Rdb_dict_manager::get_index_info(
 
   if (error) {
     rdb_fatal_error(
-        "RocksDB: Found invalid key version number (%u, %u, %u, %llu) "
+        "RocksDB: Found invalid key version number (%u, %u, %u, %" PRIu64
+        ") "
         "from data dictionary. This should never happen "
         "and it may be a bug.",
-        index_info->m_index_dict_version, index_info->m_index_type,
-        index_info->m_kv_version, index_info->m_ttl_duration);
+        (unsigned int)index_info->m_index_dict_version,
+        (unsigned int)index_info->m_index_type,
+        (unsigned int)index_info->m_kv_version, index_info->m_ttl_duration);
   }
 
   return found;
